requirements.txt
configparser==3.5.0
python-telegram-bot==5.0.0
ts3==1.0.3
